[zsh/keybindings] Removing git keybindings
They are not that useful given that it's the same number of keystrokes
as the aliases + return but they are less ergonomic to type so I am
removing them. On the other hand, the other ones [,f|,c|,gb] are very
much useful.

Modified the switch branch for something more ergo